Use translation-server v2
This also removes support for APIv1 web translation. Web translation
isn't a public feature, and I think the APIv1 version was only used by
an old version of the bookmarklet and only by the Amazon translator. The
old version of the bookmarklet we're still serving to some browsers uses
APIv2.