Add /file/view/url to return view URL without redirect
This is necessary for the new JS-based web library to send users to a
pre-signed, time-limited file URL. /file/view redirects to the file URL,
but it requires an API key, which isn't available to a normal web
request. The web library has the API key, but it can't get the final URL
from a redirect without following the redirect. With this endpoint, the
web library can get a pre-signed URL when the user clicks the attachment
link and send the user there.