Return 400 if <5.0.78 client tries to download object using newer schema
**Currently in log-only mode**

Before 5.0.78, unknown or invalid-for-type properties were skipped when
saving downloaded objects, so any data in such properties would be
silently discarded if we started returning it. Starting in 5.0.78,
objects with such properties trigger an error saying to upgrade to a new
version to continue syncing all data.

This change causes us to return a 400 if pre-5.0.78 client tries to
download an object using newer properties, which prevents us from having
to cut off those older clients from syncing altogether.