Remove 'yesterday'/'today'/'tomorrow' parsing from strToDate()
Similar to zotero/zotero@a549a64de9, this probably shouldn't have ever
been here. In this case, it meant that a value of 'today' in a date
field would be converted to "2020-04-16 today" as the visible (not
multipart) date value, because 'today' would be converted to a multipart
string with today's date but then it would fail the `isMultipart()` test
(since the year didn't appear after the SQL date) and just be used as
the real value.

This parsing should just be handled by the API client.

Addresses zotero/web-library#366