Always roll back transactions on rollback(), regardless of nesting level
The previous logic, attempting to flag nested transactions for later
rollback, could result in dangerous situations.

rollback() now rolls back all open transactions and resets the
transaction level, and a rollback() when already at the top nesting
level is just a no-op. Code calling rollback() should always throw to
indicate to calling code that now further work should be done in the
transaction. A commit() outside of a transaction throws an error.