Allow changing parent of annotation
"Convert Linked Files to Stored Files" in the client changes the parent
to a new attachment item, which wasn't allowed by the API. We could
instead always create new copies of annotations (for that function and
when merging attachments in the future), but it's less efficient and
would require recreating cached images for image annotations. It would
also be hard to fix existing annotations that have had their parents
changed. So we'll just allow it.

Also:

- Clean up changed-parent checks during item updates