Fetch 150 results from Elasticsearch for full-text search
ES defaults to 10, so previously an 'everything' search could easily
miss results. It still can if there are more than 150 full-text matches,
but it's much less likely. There might not be a problem with going
higher, but let's wait to see if anyone complains.

Unfortunately since we're not doing the entire search through ES, we
can't just use the dataserver's own pagination and instead need to fetch
all results for the full-text search to combine with the MySQL results.