Optimize username/name handling [DB update]
Avoid UPDATE on the `users` table to update the username unless the
username has actually changed, and remove unused `joined`/`lastSyncTime`
columns.

Also use the user's real name instead of the username in most API
responses.  The real name comes from the WWW DB, so fall back to the
username on master if unavailable. It would be nice not to cache the
username on master, but we want to avoid making the WWW DB a hard
dependency except when absolutely necessary (e.g., creating a new API
key).