use sqs to send reindex event to lambda
Instead of triggering the reindex lambda directly, send an event to
special queue, from where lambda will fetch those events.

The reasoning is: if reindex lambda fails and not all items are added
to DLQ, leftover items will never be indexed. But this way, event will
be retried until all items are added to DLQ.