check reindexing status + better reindexing logic
- GET request to /fulltext/reindex will return json with
reindexingStatus, which can be 'none', 'requested', 'in_progress',
or 'complete'. In case if reindexing is complete but the ES is empty,
returned reindexingStatus is 'none'.
- POST request to /fulltext/reindex will only work if the ES has 0
indexed records for that library, and if the current reindexing status
is either 'none' or 'complete'. Reindexing status file will be deleted
before sending a new reindexing request to SQS.

In both cases, reindexing status file can exist with the ES being empty
in case if a library was deleted from ES, re-indexed, then deleted
again, and another re-indexing request is submitted. Unlikely but
not impossible over the span of multiple years.