enable reindexing of library to ES
- added boolean column deindexed_from_es to the library table.
It should be set to true when the library is deleted from ES.
- GET /fulltext/reindex returns reindexingStatus that's one of
indexed, indexing or deindexed. 'deindexed' if deindexed_from_es is true.
'indexed' if # of indexed items = # of indexable attachments, and
'indexing' otherwise.
- POST /fulltext/reindex adds an event to SQS that triggers full-text-indexer
lambda that handles reindeixing and sets deindexed_from_es = false.
Only possible if deindexed_from_es is initially true.