Stop updating `keys.lastUsed`
We were updating `keys.lastUsed` for key-based requests because we
didn't always have an IP address to log to `keyAccessLog`, but now that
we get an IP address from all internal services (streaming server, www
groups requests) via `Zotero-Forwarded-For`, just log to `keyAccessLog`
and get the most recent timestamp from there.

We're continuing to check `lastUsed` as well just because it can have a
more recent value (e.g., if someone only connected to the streaming
server and didn't sync), but we'll stop worrying about that and remove
`lastUsed` once recent requests are all logged in `keyAccessLog`.