Don't return empty fields from newer schema to old clients
Check archived versions when generating JSON for a request with an old
`Zotero-Schema-Version` and omit empty fields if they weren't present in
that schema version (or the next known schema version above it).
Populated fields will still be returned and trigger unknown-field errors
in Zotero clients, but this avoids triggering errors if an old client is
retrieving data that doesn't actually contain new fields or making a
write.

This also removes some DB calls for item-field checks in favor of the
local JSON file.