remove bidirectional nature of related items
Related items have been treated as bidirectional after https://github.com/zotero/dataserver/commit/fbaa602bdd32b0f43b8997f51aa13686986eab01.
This means that if item A is one of related items for item B but not the other way around,
item B would still be included in the list of related items for item A.

This caused sync issue when one of two related items was deleted.
Modifications to the non-deleted item are synced first,
followed by the deletion.
The dataserver would return the relation to the item that was about
to be deleted after modification of the non-deleted item is handled.
This would trick Zotero on the client side into believing that
the relation to the deleted item still existed.

More context: https://github.com/zotero/zotero/pull/5486